pyproject.toml
[project]
name = "chat-teacher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain[openai]>=0.3.25",
    "pydantic>=2.11.6",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
